Convertido el proyecto para que sea un paquete de composer.
Además se ha agregado una carpeta tests con un ejemplo de su uso.
