+++ .gitignore
@@ -1,0 +1,2 @@
+/vendor/
+tests/salida.xsig
