+++ composer.json
@@ -1,0 +1,21 @@
+{
+    "name": "josemmo/facturae-php",
+    "description": "Librería para la generación y firma de facturas Facturae.",
+    "license": "MIT",
+    "homepage": "https://github.com/josemmo/Facturae-PHP",
+    "authors": [
+        {
+            "name": "josemmo",
+            "email": "josemmo@protonmail.com"
+        }
+    ],
+    "require": {
+        "php": ">=5.6",
+        "ext-openssl": "*"
+    },
+    "autoload": {
+        "psr-4": {
+            "josemmo\\Facturae\\": "src/"
+        }
+    }
+}
