+++ tests/public.pem
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEMTCCAxmgAwIBAgIJAKeVepKRtaTpMA0GCSqGSIb3DQEBCwUAMIGuMQswCQYD
+VQQGEwJFUzEQMA4GA1UECAwHU2Vnb3ZpYTEQMA4GA1UEBwwHU2Vnb3ZpYTEbMBkG
+A1UECgwSRmFsc2UgY29tcGFueSBTLkwuMSAwHgYDVQQLDBdGYWxzZSBPcmdhbml6
+YXRpb24gVW5pdDEaMBgGA1UEAwwRRmFsc2UgQ29tbW9uIE5hbWUxIDAeBgkqhkiG
+9w0BCQEWEWZhbHNlQGV4YW1wbGUuY29tMB4XDTE3MDkwNzE1MjU0MFoXDTE4MDkw
+NzE1MjU0MFowga4xCzAJBgNVBAYTAkVTMRAwDgYDVQQIDAdTZWdvdmlhMRAwDgYD
+VQQHDAdTZWdvdmlhMRswGQYDVQQKDBJGYWxzZSBjb21wYW55IFMuTC4xIDAeBgNV
+BAsMF0ZhbHNlIE9yZ2FuaXphdGlvbiBVbml0MRowGAYDVQQDDBFGYWxzZSBDb21t
+b24gTmFtZTEgMB4GCSqGSIb3DQEJARYRZmFsc2VAZXhhbXBsZS5jb20wggEiMA0G
+CSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDei4mMER6QAB0wenpbQmSnkpCbHyFE
+8X3h/0XIMsBAkAdMfzI+LEoRJ07d3IJRHsCmR0D5+TFsXJERLATKu6MO8dtGSr6m
+vnGr8jI+iZrc/E7pcTjXsYuLTp6wNNthWoDjShFDbL4LBXnsGwl2kNEFh3JE3ER+
+xqczUnw+MK2uSwfaSFu2shafkH2UAHXEnutm+cjFLW76qKlzdl9dIRublgz1EFMP
+nmp3IKbiSzAV8LN6SBhD2l/MnY0BKm5dE0pi60h9bvOWTfnupPNTl96wueWTMgvM
+1RGmIe0f6CsrDLfVJ0ytbf2/p2beat+9dbmfUocKYD1W50rsUcB6hIk5AgMBAAGj
+UDBOMB0GA1UdDgQWBBTnigCZhH8V3DVgaGBKzqMzkXVZajAfBgNVHSMEGDAWgBTn
+igCZhH8V3DVgaGBKzqMzkXVZajAMBgNVHRMEBTADAQH/MA0GCSqGSIb3DQEBCwUA
+A4IBAQAnDSdnKoj8DHLsQPD5Oypl8AdopB9VZrFKIxiK7ZK5mcfDJMQ9vf/Kb+xU
+E6GxJo68ynaOQ37tYdDkMRVdnms5oy0k+wMTAMUc0FynORP5sTdHwmrO69/HOw/C
+PjuM3+iEIPCxjP5UwRa4XnjzwNtePDv7R3JnihkLw+eSm0+nm+K0VKZ4hxvJvIgC
+YX5E+/aezHUcMvXgV0+hPR3oYQFY20Yde4XGP54L95QCHR9jbX5KCHWLs9hizcH7
+M1lpEioEj3X9R0STz5Viw5wHc0N7BAC1iZd1G3T+aK/WejZvO1/lxzghEku4b/SK
+RTQne68i20CYdE8lABGspUlFUKp5
+-----END CERTIFICATE-----
